Fixes for idSpring (func_spring) from FraggingFree by IvanTheB, fix #31
In addition to Ivans code I made two small changes:
- Add "compress" and "pullEnt1" spawnargs so those can be set when
  initializing the underlying idForce_Spring
- Made "pullEnt1" true by default (in FraggingFree it's hardcoded
  to false)

"pullEnt1" specifies whether the entity connected through "ent1" can be
moved by the spring - if it's set to 0, only ent2 can be moved by the
spring and ent1 is either stationary or moved by other means/forces,
so with "pullEnt1" "0", ent1 drags ent2 along, but not the other way
around.
"compress" is like "constant" (aka Kstretch), but for compressing when
below the "restlength" instead of stretching when above it.

# Conflicts:
#	neo/d3xp/Misc.cpp
#	neo/d3xp/Misc.h
#	neo/d3xp/physics/Force_Spring.cpp
#	neo/game/Misc.cpp
#	neo/game/Misc.h
#	neo/game/physics/Force_Spring.cpp