Fix crashes with missing MD3 models
Part of the problem was that (on 64bit?)
`if ( size <= sizeof(md3Header_t) )`
doesn't work as expected because size is a (signed) int(32),
while sizeof(...) returns an (unsigned) size_t and apparently
int size = -1 is casted to size_t then and becomes really big...

Another part is that the standard way to detect (from the "outside")
whether the model could be loaded is `model->IsDefaultModel()` and
for that to return true one must call `MakeDefaultModel()`.

I'm not sure why I need to treat this specially in
`idRenderModelManagerLocal::GetModel()`, but if I don't,
`idRenderModelMD3::InstantiateDynamicModel()` is called all the time
later and crashes (at least unless I also add a special case there).
But this problem should be the same for other kinds of models?
(Didn't test missing MD5 or ASE or whatever models though)

refs #698